ajouter les buttons de suppression et modification

+++ app/src/main/java/com/example/shop/Models/Produits.java
@@ -68,7 +68,7 @@ public class Produits {
         return prix;
     }
 
-    public void setPrix(float prix) {
+    public void setPrix(double prix) {
         this.prix = (double) prix;
     }
